test18

+++ header/he.h
@@ -1,7 +1,7 @@
 #include<stdio.h>
 #include<stdlib.h>
 #include<pthread.h>
-void push(char *w,int len,int coun);
+void push(char *w,int len,long int coun);
 int search(char *w,int len);
 void *ficount(char *c1);
 pthread_mutex_t mutex;
